All errors should be fixed
I'm pretty sure the errors we were getting were somewhere in statement(). If they weren't, then I have no idea what could possibly be wrong. The only real difference in code between this and parser.c is the lack of macros in this one and having different names for a few functions.